initialize empty dtype
I did not know the empty default type is float, it is more efficient to initialize it with the same dtype of the image

+++ skimage/feature/blob.py
@@ -360,7 +360,7 @@ def blob_dog(image, min_sigma=1, max_sigma=50, sigma_ratio=1.6, threshold=0.5,
     # computing difference between two successive Gaussian blurred images
     # to obtain an approximation of the scale invariant Laplacian of the
     # Gaussian operator
-    image_cube = np.empty(shape=image.shape + (k,))
+    image_cube = np.empty(shape=image.shape + (k,), dtype=float_dtype)
     for i in range(k):
         image_cube[..., i] = (gaussian_images[i] - gaussian_images[i + 1]) * sf
 
@@ -525,7 +525,7 @@ def blob_log(image, min_sigma=1, max_sigma=50, num_sigma=10, threshold=.2,
 
     # computing gaussian laplace
     # average s**2 provides scale invariance
-    image_cube = np.empty(shape=image.shape + (len(sigma_list),))
+    image_cube = np.empty(shape=image.shape + (len(sigma_list),), dtype=float_dtype)
     for j, s in enumerate(sigma_list):
         image_cube[..., j] = -ndi.gaussian_laplace(image, s) * np.mean(s) ** 2
 
@@ -665,7 +665,7 @@ def blob_doh(image, min_sigma=1, max_sigma=30, num_sigma=10, threshold=0.01,
     else:
         sigma_list = np.linspace(min_sigma, max_sigma, num_sigma)
 
-    image_cube = np.empty(shape=image.shape + (len(sigma_list),))
+    image_cube = np.empty(shape=image.shape + (len(sigma_list),), dtype=float_dtype)
     for j, s in enumerate(sigma_list):
         image_cube[..., j] = _hessian_matrix_det(image, s)
